fix: deadlock at starting up
