+++ src/store/useAirdropStore.ts
@@ -209,7 +209,7 @@ export const setAirdropTokens = async (airdropTokens?: AirdropTokens) => {
 
         console.log({ airdropTokens });
         await invoke('set_airdrop_tokens', {
-            airdropTokens: airdropTokens,
+            airdropTokens: { token: airdropTokens.token, refresh_token: airdropTokens.refreshToken },
         });
     } else {
         // User not connected
